+++ MVC/MvcSolar/Controllers/UtilizadoresController.cs
@@ -117,7 +117,7 @@ namespace MvcSolar.Controllers
         }
 
         // GET: Utilizadores/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError=false)
         {
             if (id == null)
             {
@@ -125,10 +125,16 @@ namespace MvcSolar.Controllers
             }
 
             var utilizador = await _context.Utilizadores
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.UtilizadorId == id);
             if (utilizador == null)
             {
                 return NotFound();
+            }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Apagar falhou. Tente outra vez, e se o problema persistir, contacte o administrador.";
             }
 
             return View(utilizador);
@@ -140,10 +146,21 @@ namespace MvcSolar.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var utilizador = await _context.Utilizadores.FindAsync(id);
-            _context.Utilizadores.Remove(utilizador);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+            if (utilizador == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Utilizadores.Remove(utilizador);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
+         }
 
         private bool UtilizadorExists(int id)
         {
